Забыл подключить файл, мне кажется мне нужно спатьььь

+++ taskaev.matvei/S5/main.cpp
@@ -2,6 +2,7 @@
 #include <fstream>
 #include <string>
 #include <functional>
+#include "functions.hpp"
 #include "BSTree.hpp"
 
 int main(int argc, char ** argv)
@@ -12,7 +13,7 @@ int main(int argc, char ** argv)
     std::cerr << "Error: not open file!\n";
     return 1;
   }
-  BSTree< size_t, std::string> tree;
+  BSTree< int, std::string> tree;
   std::ifstream inputs(argv[1]);
   if (!inputs)
   {
